Update fft lib, add verif for every lib, fix nstreams

+++ gpu4s_benchmark/fast_fourier_transform_bench/main.cpp
@@ -99,7 +99,8 @@ int main(int argc, char *argv[]){
 	// copy memory to device
 	copy_memory_to_device(fft_bench, d_B, size_B);
 	// execute kernel
-	execute_kernel(fft_bench, arguments_parameters->size>>1);
+	//FIX: cast the size to select the good protytpe
+	execute_kernel(fft_bench, (int64_t)arguments_parameters->size>>1);
 	// copy memory to host
 	copy_memory_to_host(fft_bench, d_B, size_B);
